move to project_user_details endpoint and refactor
