+++ src/sentry/api/endpoints/project_users.py
@@ -3,7 +3,6 @@ from rest_framework.response import Response
 from sentry.api.bases.project import ProjectEndpoint
 from sentry.api.paginator import DateTimePaginator
 from sentry.api.serializers import serialize
-from sentry.auth.superuser import is_active_superuser
 from sentry.models import EventUser
 
 
@@ -43,30 +42,3 @@ class ProjectUsersEndpoint(ProjectEndpoint):
             paginator_cls=DateTimePaginator,
             on_results=lambda x: serialize(x, request.user),
         )
-
-    def delete(self, request, project):
-        """
-        Delete an Event User
-        ````````````````````````````````
-
-        Delete an event's user.
-
-        :pparam string organization_slug: the slug of the organization.
-        :pparam string project_slug: the slug of the project.
-        :qparam string query: Prefix ``id`` should be used to match on.
-                              For example, ``query=id:3``
-        """
-        if is_active_superuser(request):
-            identifier = [v for k, v in request.query_params.items()]
-            pieces = identifier[0].strip().split(":", 1)
-            queryset = EventUser.objects.filter(project_id=project.id, ident=pieces[1])
-
-            queryset[0].delete()
-
-            return self.paginate(
-                request=request,
-                queryset=queryset,
-                order_by="-date_added",
-                paginator_cls=DateTimePaginator,
-                on_results=lambda x: serialize(x, request.user),
-            )
